Moved fields export logic to factory classes
